Renames database mod to database_setup.
