+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod app;
-pub mod database;
+pub mod database_setup;
 pub mod intro;
 pub mod voice;
